+++ service/database/db-structs.go
@@ -5,66 +5,66 @@ type DatabaseLogin struct {
 }
 
 func DatabaseLoginDefault() DatabaseLogin {
-	return DatabaseLogin {
+	return DatabaseLogin{
 		Username: "",
 	}
 }
 
 type DatabaseUser struct {
-	Id uint64 `json:"id"`
+	Id       uint32 `json:"id"`
 	Username string `json:"username"`
 }
 
 func DatabaseUserDefault() DatabaseUser {
-	return DatabaseUser {
-		Id: 0,
+	return DatabaseUser{
+		Id:       0,
 		Username: "",
 	}
 }
 
 type DatabasePhoto struct {
-	Id uint64 `json:"id"`
-	User DatabaseUser `json:"user"`
-	Url string `json:"url"`
-	Date string `json:"date"`
-	LikeCount int `json:"like_count"`
-	CommentCount int `json:"comment_count"`
+	Id           uint32       `json:"id"`
+	User         DatabaseUser `json:"user"`
+	Url          string       `json:"url"`
+	Date         string       `json:"date"`
+	LikeCount    int          `json:"like_count"`
+	CommentCount int          `json:"comment_count"`
 }
 
 func DatabasePhotoDefault() DatabasePhoto {
-	return DatabasePhoto {
-		Id: 0,
-		User: DatabaseUserDefault(),
-		Url: "",
-		Date: "",
-		LikeCount: 0,
+	return DatabasePhoto{
+		Id:           0,
+		User:         DatabaseUserDefault(),
+		Url:          "",
+		Date:         "",
+		LikeCount:    0,
 		CommentCount: 0,
 	}
 }
 
 type DatabaseComment struct {
-	Id uint64 `json:"id"`
+	Id          uint32 `json:"id"`
 	CommentBody string `json:"comment_body"`
 }
 
 func DatabaseCommentDefault() DatabaseComment {
-	return DatabaseComment {
-		Id: 0,
+	return DatabaseComment{
+		Id:          0,
 		CommentBody: "",
 	}
 }
 
 type DatabaseProfile struct {
-	User DatabaseUser `json:"user"`
-	PhotosCount int `json:"photos_count"`
-	FollowersCount int `json:"followers_count"`
-	FollowingCount int `json:"following_count"`
+	User           DatabaseUser `json:"user"`
+	PhotosCount    int          `json:"photos_count"`
+	FollowersCount int          `json:"followers_count"`
+	FollowingCount int          `json:"following_count"`
 }
 
 func DatabaseProfileDefault() DatabaseProfile {
-	return DatabaseProfile {
-		User: DatabaseUserDefault(),
-		PhotosCount: 0,
+	return DatabaseProfile{
+		User:           DatabaseUserDefault(),
+		PhotosCount:    0,
 		FollowersCount: 0,
 		FollowingCount: 0,
 	}
@@ -77,7 +77,7 @@ type DatabaseStream struct {
 func DatabaseStreamDefault() DatabaseStream {
 	emptyArray := [0]DatabasePhoto{}
 
-	return DatabaseStream {
+	return DatabaseStream{
 		Photos: emptyArray[:],
 	}
 }
@@ -89,7 +89,7 @@ type DatabaseUserList struct {
 func DatabaseUserListDefault() DatabaseUserList {
 	emptyArray := [0]DatabaseUser{}
 
-	return DatabaseUserList {
+	return DatabaseUserList{
 		Users: emptyArray[:],
 	}
 }
@@ -101,7 +101,7 @@ type DatabaseCommentList struct {
 func DatabaseCommentListDefault() DatabaseCommentList {
 	emptyArray := [0]DatabaseComment{}
 
-	return DatabaseCommentList {
+	return DatabaseCommentList{
 		Comments: emptyArray[:],
 	}
 }
